Move TelemetryEventsCollector and EndpointTelemetryFilter to Telemetry namespace
