+++ application/account-management/Tests/EndpointBaseTest.cs
@@ -8,8 +8,8 @@ using PlatformPlatform.SharedKernel.Authentication;
 using PlatformPlatform.SharedKernel.ExecutionContext;
 using PlatformPlatform.SharedKernel.Integrations.Email;
 using PlatformPlatform.SharedKernel.SinglePageApp;
-using PlatformPlatform.SharedKernel.TelemetryEvents;
-using PlatformPlatform.SharedKernel.Tests.TelemetryEvents;
+using PlatformPlatform.SharedKernel.Telemetry;
+using PlatformPlatform.SharedKernel.Tests.Telemetry;
 
 namespace PlatformPlatform.AccountManagement.Tests;
