+++ application/back-office/Tests/EndpointBaseTest.cs
@@ -6,8 +6,8 @@ using Microsoft.Extensions.DependencyInjection;
 using PlatformPlatform.SharedKernel.ExecutionContext;
 using PlatformPlatform.SharedKernel.Integrations.Email;
 using PlatformPlatform.SharedKernel.SinglePageApp;
-using PlatformPlatform.SharedKernel.TelemetryEvents;
-using PlatformPlatform.SharedKernel.Tests.TelemetryEvents;
+using PlatformPlatform.SharedKernel.Telemetry;
+using PlatformPlatform.SharedKernel.Tests.Telemetry;
 
 namespace PlatformPlatform.BackOffice.Tests;
